Disallow duplicates when using addEdge on a graphNEL
Up until now, one could add duplicate edges to a graphNEL using
addEdge.  This can complicated reading serialized graphs, such as
GXL, because for an undirected graph it is reasonable to have the
reciprocal edges in the serialization (or not).

This change does not add anything to the validity check, so a
valid graphNEL can still contain dup edges, but it is slightly
more difficult to create one now.

In the future, we plan to implement a multiGraphNEL or some such
for the purpose of representing graphs with multiple edges
between a given pair of nodes.


git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/graph@16434 bc3139a8-67e5-0310-9ffc-ced21a209358