Use getExportedValue instead of require where possible for XML package
For functions that provide optional features based on the XML package,
we now use getExportedValue instead of require("XML") to avoid
attaching the XML package to the search path and polluting the user's
visible namespace.  An exception is the toGXL function which returns
an XML object.  This function still calls require.


git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/graph@27073 bc3139a8-67e5-0310-9ffc-ced21a209358